+++ CHANGES.rst
@@ -7,8 +7,9 @@ Version 1.x is for PloneMeeting 4.1.x, version 2.x is for PloneMeeting 4.2.x+
 2.5 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- When using `fullobjects`, only serialize `next/previous` when specifically
+  asked using `include_nextprev=true`.
+  [gbastien]
 
 2.4 (2024-03-14)
 ----------------